refactor: get rid of types.h file for enums
